Moved AdminNicety into its own file
